requirements.txt
pytesseract>=0.3.10
Pillow>=9.0.0
reportlab>=3.6.0
